Added validations

+++ src/main/java/com/aliyaa/assignment/mmt/flightDetails/cont/FlightDetailsController.java
@@ -1,5 +1,7 @@
 package com.aliyaa.assignment.mmt.flightDetails.cont;
 
+import java.io.IOException;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -7,6 +9,9 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,6 +25,7 @@ import com.aliyaa.assignment.mmt.flightDetails.DTO.FlightsDTO;
 import com.aliyaa.assignment.mmt.flightDetails.entity.FareDetails;
 //import com.aliyaa.assignment.mmt.flightDetails.entity.FareDetails;
 import com.aliyaa.assignment.mmt.flightDetails.entity.Flights;
+import com.aliyaa.assignment.mmt.flightDetails.exceptions.Validations;
 import com.aliyaa.assignment.mmt.flightDetails.service.FlightDetailsService;
 
 @RestController
@@ -27,6 +33,8 @@ public class FlightDetailsController {
 	
 	@Autowired(required = false)   
 	private FlightDetailsService flightService;    
+	
+	Validations validations =new Validations();
 	
 	public FlightDetailsService getFlightService() {
 		return flightService;
@@ -72,11 +80,20 @@ public class FlightDetailsController {
 	@GetMapping("/getAllFlights/{flightNumber}")
 	public Optional<Flights> findByFlightNumber(@PathVariable int flightNumber)
 	{
-		return flightService.findById(flightNumber);
+		
+		Optional<Flights> flights=flightService.findById(flightNumber);
+	if(!flights.isPresent())
+		throw new RuntimeException("Flight number is invalid "+ flightNumber);
+	     
+		return flights;
 	}
 	@DeleteMapping("/deleteFlights/{flightNumber}")
 	public String deleteByFlightNumber(@PathVariable int flightNumber)
 	{
+		Optional<Flights> flights=flightService.findById(flightNumber);
+		if(!flights.isPresent())
+			throw new RuntimeException("Flight number is invalid "+ flightNumber);
+		     
 		flightService.deleteByFlightNumber(flightNumber);
 		return "Deleted flight number" + flightNumber;
 
@@ -87,29 +104,38 @@ public class FlightDetailsController {
 	return "Deleted All flight";
 	}
 	@GetMapping("/searchFlight")
-	public List<FlightsDTO>  findFlights(@RequestParam("source") String source, @RequestParam("destination") String destination ,
+	public ResponseEntity<List<FlightsDTO>>  findFlights(@RequestParam("source") String source, @RequestParam("destination") String destination ,
 			@RequestParam("departureDate") LocalDate departureDate, @RequestParam("classType") String classType,
-			@RequestParam Boolean roundTrip ,@RequestParam(required=false) LocalDate returnDate,@RequestParam Boolean sort,
-			@RequestParam(required=false) String sortingType,@RequestParam Boolean departure, @RequestParam(required=false) String departureType ){
-		
+			@RequestParam(required=false) String roundTrip ,@RequestParam(required=false) LocalDate returnDate, @RequestParam(required=false) String sort,
+			@RequestParam(required=false) String sortingType,@RequestParam(required=false) String departure, @RequestParam(required=false) String departureType ) throws IOException {
+		classType = classType.toLowerCase();
+		classType = StringUtils.capitalize(classType);
+		validations.classValid(classType.toLowerCase());
+			validations.trueFalse(departure.toLowerCase());
+			validations.trueFalse(sort.toLowerCase());
+			validations.trueFalse(roundTrip.toLowerCase());
+			
 		List<Flights> oneWayTripFlights=new ArrayList<>();
 		oneWayTripFlights=flightService.searchFlights(source,destination,departureDate,classType);
 
 
-		if(roundTrip==true)
+		if(roundTrip.toLowerCase().equals("true"))
 		{
+		validations.departAndArriveDate(departureDate,returnDate);
 			List<Flights> roundWayTripFlights=flightService.searchFlights(destination,source,returnDate,classType);
 			oneWayTripFlights.addAll(roundWayTripFlights);
 
 			
 		}
 		
-		if(sort==true) {
-			if(sortingType.equals("duration"))
+		if(sort.toLowerCase().equals("true")) {
+			validations.sortingTypes(sortingType.toLowerCase());
+			
+			if(sortingType.toLowerCase().equals("duration"))
 			{
 				
 				oneWayTripFlights=flightService.sortFlight1(source,destination,departureDate,classType);
-				if(roundTrip==true)
+				if(roundTrip.toLowerCase().equals("true"))
 				{
 					List<Flights> roundWayTripFlights=flightService.sortFlight1(destination,source,returnDate,classType);
 					oneWayTripFlights.addAll(roundWayTripFlights);
@@ -118,10 +144,11 @@ public class FlightDetailsController {
 				
 
 			}
-			else
+			else if(sortingType.toLowerCase().equals("fares"))
+				
 			{
 				oneWayTripFlights=flightService.sortFlight2(source,destination,departureDate,classType);
-				if(roundTrip==true)
+				if(roundTrip.toLowerCase().equals("true"))
 				{
 					List<Flights> roundWayTripFlights=flightService.sortFlight2(destination,source,returnDate,classType);
 					oneWayTripFlights.addAll(roundWayTripFlights);
@@ -131,13 +158,14 @@ public class FlightDetailsController {
 			}
 			
 		}
-		if(departure==true) {
-		
-			if(departureType.equals("morning"))
+		if(departure.toLowerCase().equals("true")) {
+			validations.filter(departureType.toLowerCase());
+			
+			if(departureType.toLowerCase().equals("morning"))
 			{
 				oneWayTripFlights=flightService.filterByMorning(source,destination,departureDate,classType);
 				//return oneWayTripFlights;
-				if(roundTrip==true)
+				if(roundTrip.toLowerCase().equals("true"))
 				{
 					List<Flights> roundWayTripFlights=flightService.filterByMorning(destination,source,returnDate,classType);
 					oneWayTripFlights.addAll(roundWayTripFlights);
@@ -145,11 +173,11 @@ public class FlightDetailsController {
 				}
 				
 			}
-			else
+			else if(departureType.toLowerCase().equals("evening"))
 			{
 				oneWayTripFlights=flightService.filterByEvening(source,destination,departureDate,classType);
 				//return oneWayTripFlights;
-				if(roundTrip==true)
+				if(roundTrip.toLowerCase().equals("true"))
 				{
 					List<Flights> roundWayTripFlights=flightService.filterByEvening(destination,source,returnDate,classType);
 					oneWayTripFlights.addAll(roundWayTripFlights);
@@ -157,8 +185,9 @@ public class FlightDetailsController {
 				}
 				
 			}
+			
 		}
-		return  oneWayTripFlights.stream().map(flights-> new FlightsDTO(
+		List<FlightsDTO> abc =  oneWayTripFlights.stream().map(flights-> new FlightsDTO(
 				flights.getAirlines(),
 				flights.getDepartureTime(),
 				flights.getArrivalTime(),
@@ -167,7 +196,7 @@ public class FlightDetailsController {
 				)).collect(Collectors.toList())
 				;
 		
-		
+		return new ResponseEntity<>(abc, HttpStatus.OK);
 
 	}
